# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: PakMart e-commerce API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /cart/items:
    post:
      operationId: addCartItem
      tags: [cart]
      summary: Add item to cart
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AddCartItemBody" }
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Cart" }
        "404":
          description: Product not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /cart/items/{id}:
    patch:
      operationId: updateCartItem
      tags: [cart]
      summary: Update cart item quantity
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateCartItemBody" }
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Cart" }
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
    delete:
      operationId: removeCartItem
      tags: [cart]
      summary: Remove cart item
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Cart" }
...
  /orders:
    post:
      operationId: createOrder
      tags: [orders]
      summary: Place order from current cart
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateOrderBody" }
      responses:
        "201":
          description: Order created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Order" }
        "400":
          description: Invalid
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get an order by id
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Order" }
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error: { type: string }
...
    CartItem:
      type: object
      required:
        - id
        - productId
        - productSlug
        - productName
        - productBrand
        - productImage
        - price
        - quantity
        - stock
        - lineTotal
      properties:
        id: { type: integer }
        productId: { type: integer }
        productSlug: { type: string }
        productName: { type: string }
        productBrand: { type: string }
        productImage: { type: string }
        price: { type: integer }
        quantity: { type: integer }
        stock: { type: integer }
        lineTotal: { type: integer }
    Cart:
      type: object
      required:
        - items
        - itemCount
        - subtotal
        - deliveryFee
        - discount
        - total
        - freeDeliveryThreshold
        - amountToFreeDelivery
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/CartItem" }
        itemCount: { type: integer }
        subtotal: { type: integer }
        deliveryFee: { type: integer }
        discount: { type: integer }
        total: { type: integer }
        freeDeliveryThreshold: { type: integer }
        amountToFreeDelivery: { type: integer }
    AddCartItemBody:
      type: object
      required: [productId]
      properties:
        productId: { type: integer }
        quantity: { type: integer }
    UpdateCartItemBody:
      type: object
      required: [quantity]
      properties:
        quantity: { type: integer }
    OrderItem:
      type: object
      required:
        - id
        - productId
        - productSlug
        - productName
        - productBrand
        - productImage
        - price
        - quantity
        - lineTotal
      properties:
        id: { type: integer }
        productId: { type: integer }
        productSlug: { type: string }
        productName: { type: string }
        productBrand: { type: string }
        productImage: { type: string }
        price: { type: integer }
        quantity: { type: integer }
        lineTotal: { type: integer }
    CreateOrderBody:
      type: object
      required: [fullName, phone, address, city, paymentMethod]
      properties:
        fullName: { type: string }
        phone: { type: string }
        email: { type: ["string", "null"] }
        address: { type: string }
        city: { type: string }
        paymentMethod:
          type: string
          enum: [cod, easypaisa, jazzcash, card]
        notes: { type: ["string", "null"] }
    Order:
      type: object
      required:
        - id
        - orderNumber
        - status
        - fullName
        - phone
        - address
        - city
        - paymentMethod
        - subtotal
        - deliveryFee
        - discount
        - total
        - items
        - createdAt
      properties:
        id: { type: integer }
        orderNumber: { type: string }
        status:
          type: string
          enum: [pending, confirmed, shipped, delivered, cancelled]
        fullName: { type: string }
        phone: { type: string }
        email: { type: ["string", "null"] }
        address: { type: string }
        city: { type: string }
        paymentMethod: { type: string }
        notes: { type: ["string", "null"] }
        subtotal: { type: integer }
        deliveryFee: { type: integer }
        discount: { type: integer }
        total: { type: integer }
        items:
          type: array
          items: { $ref: "#/components/schemas/OrderItem" }
        createdAt:
          type: string
          format: date-time